requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1

# --- LangChain (Pinned to a known-good, modern, and stable set) ---
langchain==0.2.1
langchain-core==0.2.5
langchain-community==0.2.1
langchain-text-splitters==0.2.0
langchain-google-genai==1.0.5
langchain-huggingface==0.0.3

# --- Core RAG and AI Components ---
faiss-cpu==1.8.0
sentence-transformers==2.7.0
pypdf==4.2.0

# --- Google Specific Dependency ---
google-generativeai==0.5.4